french-vocab-bot/requirements.txt
python-telegram-bot==20.7
aiosqlite
python-dotenv
APScheduler
